Refact GraphPage by creating a Graph component
